Allow OBJDIR cleaning to be overridden.
Cleaning `OBJDIR` can take a long time and, in the interactive case, may
do more harm than good (in that people will avoid this script out of
frustration, reducing burn-in by the project). Allow this scrubbing to
be overridden by setting the `NO_CLEAN` variable.

+++ scripts/cadets-build.sh
@@ -88,8 +88,13 @@ echo "------------------------------------------------------------------------"
 # Clean up old obj tree but don't delete any package repositories.
 make -C ${SRCDIR} obj
 OBJDIR=`make -C ${SRCDIR} -V .OBJDIR`
-find ${OBJDIR} -depth 1 -not -name repo | xargs rm -rf
 
+if [ "${NO_CLEAN}" = "" ]
+then
+	echo "Cleaning up OBJDIR: ${OBJDIR}"
+	find ${OBJDIR} -depth 1 -not -name repo | xargs rm -rf
+	echo "OBJDIR clean"
+fi
 
 
 # Disable -Werror in both buildworld (NO_WERROR) and buildkernel (WERROR),
